test_images/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
psycopg2-binary==2.8.6
gunicorn==20.1.0
